DDF-2076: Use a RNG for the generation fo SessionIndex in the SAML element AuthnStatement   - Created a RandomNumberGenerator class to share a common RNG that uses the BC seed
