+++ platform/util/platform-util/pom.xml
@@ -40,6 +40,11 @@
             <artifactId>org.apache.felix.configadmin</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.bouncycastle</groupId>
+            <artifactId>bcprov-jdk15on</artifactId>
+            <version>${bouncy.version}</version>
+        </dependency>
     </dependencies>
     <build>
         <plugins>
